requirements.txt
setuptools
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.0
django-filter==23.5
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
gunicorn==21.2.0
whitenoise==6.6.0


